requirements.txt
asgiref==3.6.0
Django==4.1.6
django-countries==7.5.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
phonenumbers==8.13.5
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
